+++ Mindstorms.Core/Commands/Command.cs
@@ -12,13 +12,13 @@ namespace Mindstorms.Core.Commands
             get { return data.ToArray(); }
         }
 
-        protected readonly List<byte> SystemCommandNoReply = new List<byte> { (byte)CommandType.SystemCommand | (byte)Response.NotExpected };
+        protected readonly List<byte> SystemCommandNoReply = new List<byte> { CommandType.SystemCommand | Response.NotExpected };
 
-        protected readonly List<byte> SystemCommandWithReply = new List<byte> { (byte)CommandType.SystemCommand | (byte)Response.Required };
+        protected readonly List<byte> SystemCommandWithReply = new List<byte> { CommandType.SystemCommand | Response.Required };
 
-        protected readonly List<byte> DirectCommandNoReply = new List<byte> { (byte)CommandType.DirectCommand | (byte)Response.NotExpected, 0, 0 };
+        protected readonly List<byte> DirectCommandNoReply = new List<byte> { CommandType.DirectCommand | Response.NotExpected, 0, 0 };
 
-        protected readonly byte DirectCommandWithReply = (byte)CommandType.DirectCommand | (byte)Response.Required;
+        protected readonly byte DirectCommandWithReply = CommandType.DirectCommand | Response.Required;
 
         protected List<byte> GetDirectCommandWithReply(byte numberOfExpectedBytes, byte localBytes = 0) => new List<byte> { DirectCommandWithReply, numberOfExpectedBytes, (byte)(localBytes << 2) };
     }
