+++ Mindstorms.Core/Commands/File/DownloadFileFromBrick.cs
@@ -11,7 +11,7 @@ namespace Mindstorms.Core.Commands.File
             var bytesToRead = (ushort)Math.Min(fileSize, Constants.ChunkSize);
 
             data = SystemCommandWithReply;
-            data.Add((byte)SystemCommand.BeginFileUpload);
+            data.Add(SystemCommand.BeginFileUpload);
             data.Append(bytesToRead);
             data.Append(filePath);
         }
