+++ Mindstorms.Core/Commands/LCD/ChangeFontType.cs
@@ -9,9 +9,9 @@ namespace Mindstorms.Core.Commands.LCD
         public ChangeFontType(FontType fontType)
         {
             data = DirectCommandNoReply;
-            data.Add((byte)OpCode.DrawUI);
-            data.Add((byte)DrawSubCode.SelectFont);
-            data.Add((byte)fontType);
+            data.Add(OpCode.DrawUI);
+            data.Add(DrawSubCode.SelectFont);
+            data.Add(fontType);
         }
     }
 }
